Change the binary packet output order a bit so that the imu packet is the last of the 'regular' or 'every frame' data packets.  This can serve as a fairly regular time sync marker for a host system if it wishes to sync off this data packet.
The GPS packet is irregular so that is sent after the IMU whenever new gps
data is available.  The consequence to a host system syncing off the imu
packet is that the gps data will arrive one frame (1/100th of a second) late.
This shouldn't be a big issue for most systems.