Revert "Change the binary packet output order a bit so that the imu packet is the"
This reverts commit f47f316024b0c622cc31248e335ced4adae722c6.

I believe the change was too hasty and needs to be thought about some more.
I was trying for a consistent uart data transfer amount, so moved the irregular
gps packet after the imu packet which the host can key off of.  However, this
just pushes the extra uart work to the host's next processing frame and
thus doesn't really accomplish the goal.  Plus a few bytes over the uart
shouldn't consume 1-2 milleseconds which is the variation I was seeing at the
host and trying to address.  This can't be the source.